### 2019-05-02 #1422 @kyleniemeyer
@smk508 I am looking for a JOSS editor to handle the submission; once someone is assigned, they will work to assign reviewers.
