### 2019-05-03 #1422 @arokem
@smk508 : do you have any suggestions for reviewers? You can also look [here](https://bit.ly/joss-reviewers)
