### 2019-05-04 #1422 @smk508
Hi @arokem . Thanks so much! I looked through the spreadsheet for people who are experts in machine learning, big data, and/or bioinformatics. There's a lot of people, so I'm not sure who makes the most sense, but I just picked out some names. I think if there's someone who has experience designing or contributing to machine learning / big data frameworks such a TensorFlow / PyTorch / Spark, that person would probably have good insight to give since this library is a framework meant to facilitate machine learning.

Here are some names as I skimmed the spreadsheet:

Simon Dirmier
Niranjan Padmanabhan (username: neurons)
Divya Sardanda
