### 2019-05-06 #1422 @arokem
👋 @dvysardana: would you be willing to review this paper for JOSS? It seems right up your alley!
