### 2019-05-20 6aa22f2
Updated ignite trainer to allow for specifying visdom environment.
