+++ setup.py
@@ -5,7 +5,7 @@ with open('requirements.txt') as f:
 
 setup(
     name='Fireworks-ml',
-    version='0.3.0',
+    version='0.3.1',
     packages=find_packages(),
     author_email="skhan8@mail.einstein.yu.edu",
     description="A batch-processing framework for data analysis and machine learning using PyTorch.",
