+++ requirements.txt
@@ -8,5 +8,4 @@ py>=1.5.3
 pytorch-ignite>=0.1.0
 SQLAlchemy>=1.3.1
 SQLAlchemy-Utils==0.33.5
-torch>=1.0.1.post2
 torchfile>=0.1.0
