### 2019-05-20 daabc8f
Fixed typo and updated finnicky test for normalizer class
