### 2019-05-20 6a8bc66
Updated statistical tests to be more lenient
