+++ Fireworks/core/model_test.py
@@ -134,8 +134,8 @@ def test_Model_save_load():
     assert (newone.m == 5.).all()
     # old = newone.state_dict()['d.bias'].clone().detach().numpy()
     newone.load_state('save_test/torch_babura-test.json')
-    assert (newone.d.bias - babura.d.bias < .001).all()
-    assert (newone.d.weight - babura.d.weight < .001).all()
+    assert (newone.d.bias - babura.d.bias < .005).all()
+    assert (newone.d.weight - babura.d.weight < .005).all()
     # new = newone.state_dict()['d.bias'].clone().detach().numpy()
     shutil.rmtree('save_test')
 
@@ -246,11 +246,11 @@ def test_one_Model_training():
     minibatcher = get_minibatcher(training_data[0])
     train_model(A, minibatcher)
     # For some reason, this model struggles to learn the y-intercept.
-    assert (m-A.m < .4).all()
+    assert (m-A.m < .6).all()
     train_model(B, minibatcher)
-    assert (m - B.m < .4).all()
+    assert (m - B.m < .6).all()
 
-    assert (A.m - B.m < .4).all() # Test precision between models
+    assert (A.m - B.m < .6).all() # Test precision between models
 
 def test_multiple_Models_training():
     """
@@ -274,8 +274,8 @@ def test_multiple_Models_training():
     errors = training_data[1]['errors']
     minibatcher = get_minibatcher(training_data[0])
     train_model(multilinear, minibatcher)
-    assert (A.m - m1 < .4).all()
-    assert (B.m - m2 < .4).all()
+    assert (A.m - m1 < .6).all()
+    assert (B.m - m2 < .6).all()
     assert (A.b == 0).all()
     assert (C.m == 0.).all()
 
@@ -370,7 +370,7 @@ def test_multiple_Models_training_in_pipeline():
     batch.to_tensors()
     A(batch)
     train_model(B, minibatcher, models = [A, B])
-    assert (A.m - m < .4).all()
+    assert (A.m - m < .6).all()
     assert (B.b != 2).all()
     assert (B.m == 1).all()
     assert (A.b == 0).all()
@@ -423,9 +423,9 @@ def test_multiple_Models_training_via_junction():
     assert rambo
     train_model(B, minibatcher, models=[B, C, D, E])
     # Test that all Models trained
-    assert (C.m - m < .4).all()
-    assert (D.m - m < .4).all()
-    assert (E.m - m < .4).all()
+    assert (C.m - m < .6).all()
+    assert (D.m - m < .6).all()
+    assert (E.m - m < .6).all()
     assert (C.m != D.m).all()
     assert (D.m != E.m).all()
     assert (E.m != C.m).all()
