### 2019-05-21 #1422 @arokem
👋 @dirmeier : would you be willing to review this paper for JOSS?
