### 2019-05-25 #1478 @whedon
[REVIEW]: Fireworks: Reproducible Machine Learning with PyTorch
**Submitting author:** @smk508 (<a href="http://orcid.org/0000-0003-4686-4923">Saad Khan</a>)
**Repository:** <a href="https://github.com/kellylab/Fireworks" target ="_blank">https://github.com/kellylab/Fireworks</a>
**Version:** 0.3.2
