### 2019-06-02 d724289
Renamed package to 'fireworks-ml' from 'Fireworks-ml'
I am also going to add more examples and perform some more clean up.
