### 2019-06-21 56dc746
Auto stash before merge of "development" and "origin/development"
