+++ examples/VQA/utils/vqa.py
@@ -1,0 +1,180 @@
+# SOURCE: https://github.com/GT-Vision-Lab/VQA/blob/master/PythonHelperTools/vqaTools/vqa.py
+
+__author__ = 'aagrawal'
+__version__ = '0.9'
+
+# Interface for accessing the VQA dataset.
+
+# This code is based on the code written by Tsung-Yi Lin for MSCOCO Python API available at the following link: 
+# (https://github.com/pdollar/coco/blob/master/PythonAPI/pycocotools/coco.py).
+
+# The following functions are defined:
+#  VQA        - VQA class that loads VQA annotation file and prepares data structures.
+#  getQuesIds - Get question ids that satisfy given filter conditions.
+#  getImgIds  - Get image ids that satisfy given filter conditions.
+#  loadQA     - Load questions and answers with the specified question ids.
+#  showQA     - Display the specified questions and answers.
+#  loadRes    - Load result file and create result object.
+
+# Help on each function can be accessed by: "help(COCO.function)"
+
+import json
+import datetime
+import copy
+
+class VQA:
+	def __init__(self, annotation_file=None, question_file=None):
+		"""
+       	Constructor of VQA helper class for reading and visualizing questions and answers.
+        :param annotation_file (str): location of VQA annotation file
+        :return:
+		"""
+        # load dataset
+		self.dataset = {}
+		self.questions = {}
+		self.qa = {}
+		self.qqa = {}
+		self.imgToQA = {}
+		if not annotation_file == None and not question_file == None:
+			print 'loading VQA annotations and questions into memory...'
+			time_t = datetime.datetime.utcnow()
+			dataset = json.load(open(annotation_file, 'r'))
+			questions = json.load(open(question_file, 'r'))
+			print datetime.datetime.utcnow() - time_t
+			self.dataset = dataset
+			self.questions = questions
+			self.createIndex()
+
+	def createIndex(self):
+        # create index
+		print 'creating index...'
+ 		imgToQA = {ann['image_id']: [] for ann in self.dataset['annotations']}
+		qa =  {ann['question_id']:       [] for ann in self.dataset['annotations']}
+		qqa = {ann['question_id']:       [] for ann in self.dataset['annotations']}
+ 		for ann in self.dataset['annotations']:
+			imgToQA[ann['image_id']] += [ann]
+			qa[ann['question_id']] = ann
+		for ques in self.questions['questions']:
+  			qqa[ques['question_id']] = ques
+		print 'index created!'
+
+ 		# create class members
+ 		self.qa = qa
+		self.qqa = qqa
+ 		self.imgToQA = imgToQA
+
+	def info(self):
+		"""
+		Print information about the VQA annotation file.
+		:return:
+		"""
+		for key, value in self.datset['info'].items():
+			print '%s: %s'%(key, value)
+
+	def getQuesIds(self, imgIds=[], quesTypes=[], ansTypes=[]):
+		"""
+		Get question ids that satisfy given filter conditions. default skips that filter
+		:param 	imgIds    (int array)   : get question ids for given imgs
+				quesTypes (str array)   : get question ids for given question types
+				ansTypes  (str array)   : get question ids for given answer types
+		:return:    ids   (int array)   : integer array of question ids
+		"""
+		imgIds 	  = imgIds    if type(imgIds)    == list else [imgIds]
+		quesTypes = quesTypes if type(quesTypes) == list else [quesTypes]
+		ansTypes  = ansTypes  if type(ansTypes)  == list else [ansTypes]
+
+		if len(imgIds) == len(quesTypes) == len(ansTypes) == 0:
+			anns = self.dataset['annotations']
+		else:
[…]
