### 2019-06-24 263e3c8
Progress on demo for fashion MNIST
