+++ requirements.txt
@@ -1,11 +1,15 @@
-torch==0.1.1.0
-bidict>=0.17.2
-biopython>=1.71
-Deprecated>=1.2.3
-numpy>=1.14.3
-pandas>=0.23.0
-py>=1.5.3
-pytorch-ignite>=0.1.0
-SQLAlchemy>=1.3.1
+bidict==0.18.0
+biopython==1.73
+Deprecated==1.2.5
+numpy==1.16.4
+pandas==0.24.2
+py==1.8.0
+python-dateutil==2.8.0
+pytorch-ignite==0.2.0
+pytz==2019.1
+six==1.12.0
+SQLAlchemy==1.3.4
 SQLAlchemy-Utils==0.33.5
-torchfile>=0.1.0
+torch==1.1.0
+torchfile==0.1.0
+wrapt==1.11.1
