### 2019-06-25 31e8529
Finished MNIST demo
