+++ examples/mnist.py
@@ -1,9 +1,10 @@
 #%%
-from fireworks import PyTorch_Model, Message
+from fireworks import PyTorch_Model, Message, HookedPassThroughPipe, Experiment
 from fireworks.toolbox import ShufflerPipe, TensorPipe, BatchingPipe, FunctionPipe
 from fireworks.toolbox.preprocessing import train_test_split
 from fireworks.extensions import IgniteJunction
 from fireworks.core import PyTorch_Model
+import pandas as pd
 import torch
 from torchvision.datasets.mnist import FashionMNIST
 from os import environ as env
@@ -13,6 +14,8 @@ import visdom
 
 env_name = 'mnist_fashion'
 vis = visdom.Visdom(env=env_name)
+description = "Here, we will train a convolutional neural network on the Fashion MNIST dataset to demonstrate the usage of Fireworks."
+experiment = Experiment(env_name, description=description)
 #%%
 mnist_dir = env.get('MNIST_DIR', '/var/MNIST/')
 print(mnist_dir)
@@ -20,12 +23,8 @@ print(mnist_dir)
 mnist = FashionMNIST(mnist_dir, download=True)
 dataset = Message({'examples': mnist.data, 'labels': mnist.targets})
 example = dataset['examples'][0]
-#plt.imshow(example)
-#plt.show()
-
-#%% 
-# Annotate examples with class names
-classes = {i: class_name for i, class_name in zip(count(), mnist.classes)}
+plt.imshow(example)
+plt.show()
 
 #%%
 train, test = train_test_split(dataset, test=.1)
@@ -86,8 +85,8 @@ class mnistModel(PyTorch_Model):
         self.components['out_column'] = 'embeddings'
         self.components['conv1'] = torch.nn.Conv2d(1, 64, 2, padding=1)
         self.components['pool1'] = torch.nn.MaxPool2d(2)
-        #self.components['conv2'] = torch.nn.Conv2d(64, 32, 2)
-        #self.components['pool2'] = torch.nn.MaxPool2d(3)
+        self.components['conv2'] = torch.nn.Conv2d(64, 32, 2)
+        self.components['pool2'] = torch.nn.MaxPool2d(2)
         self.components['nonlinearity'] = torch.nn.ELU()
 
     def forward(self, batch):
@@ -95,10 +94,9 @@ class mnistModel(PyTorch_Model):
         embedding = batch[self.in_column]
         embedding = self.nonlinearity(self.conv1(embedding))
         embedding = self.pool1(embedding)
-        #embedding = self.nonlinearity(self.conv2(embedding))
-        #embedding = self.pool2(embedding)
-        
-        embedding = embedding.reshape(len(batch), 12544)
+        embedding = self.nonlinearity(self.conv2(embedding))
+        embedding = self.pool2(embedding)
+        embedding = embedding.reshape(len(batch), 1152)
         batch[self.out_column] = embedding
         return batch
 
@@ -109,10 +107,10 @@ class Classifier(PyTorch_Model):
     def init_default_components(self):
         self.components['in_column'] = 'embeddings'
         self.components['out_column'] = 'predictions'
-        self.components['linear1'] = torch.nn.Linear(12544, 256)
+        self.components['linear1'] = torch.nn.Linear(1152, 256)
         self.components['linear2'] = torch.nn.Linear(256, 10)
         self.components['nonlinearity'] = torch.nn.ELU()
-        self.components['softmax'] = torch.nn.Softmax()
+        self.components['softmax'] = torch.nn.Softmax(dim=1)
 
     def forward(self, batch):
         
@@ -122,13 +120,105 @@ class Classifier(PyTorch_Model):
         batch[self.out_column] = predictions
         return batch 
 
[…]
