### 2019-06-26 f6166f3
Updated mnist example to be a Jupyter notebook
