+++ full-requirements.txt
@@ -47,6 +47,7 @@ SQLAlchemy==1.3.1
 SQLAlchemy-Utils==0.33.5
 torch>=1.0.1.post2
 torchfile==0.1.0
+torchvision==0.3.0
 tornado==5.0.2
 tqdm==4.24.0
 twine==1.13.0
