+++ requirements.txt
@@ -11,5 +11,6 @@ six==1.12.0
 SQLAlchemy==1.3.4
 SQLAlchemy-Utils==0.33.5
 torch==1.1.0
+torchvision==0.3.0
 torchfile==0.1.0
 wrapt==1.11.1
