+++ requirements.txt
@@ -1,16 +1,17 @@
-bidict==0.18.0
-biopython==1.73
-Deprecated==1.2.5
-numpy==1.16.4
-pandas==0.24.2
-py==1.8.0
-python-dateutil==2.8.0
-pytorch-ignite==0.2.0
-pytz==2019.1
-six==1.12.0
-SQLAlchemy==1.3.4
-SQLAlchemy-Utils==0.33.5
-torch==1.1.0
-torchvision==0.3.0
-torchfile==0.1.0
-wrapt==1.11.1
+bidict>=0.18.0
+biopyth>==1.73
+Deprecate>=1.2.5
+numpy>=1.16.4
+pandas>0.24.2
+p>=1.8.0
+python-dateuti>=2.8.0
+pytorch-ignit>=0.2.0
+pytz>=2019.1
+six>=1.12.0
+SQLAlchem>=1.3.4
+SQLAlchemy-Utils>0.33.5
+torc>=1.1.0
+torchvisio>=0.3.0
+torchfil>=0.1.0
+wrapt=>=1.11.1
+visdom>=0.1.8.4
