+++ requirements.txt
@@ -10,7 +10,7 @@ pytz>=2019.1
 six>=1.12.0
 SQLAlchem>=1.3.4
 SQLAlchemy-Utils>0.33.5
-torc>=1.1.0
+torch>=1.1.0
 torchvisio>=0.3.0
 torchfil>=0.1.0
 wrapt=>=1.11.1
