### 2019-06-26 22faad6
update requirements and examples
