### 2019-06-26 b059996
added Binder notebook
