### 2019-06-26 520613a
Update travis badge
