+++ old-requirements.txt
@@ -1,60 +1,0 @@
-alabaster==0.7.12
-atomicwrites==1.1.5
-attrs==18.1.0
-Babel==2.6.0
-bidict==0.17.2
-biopython==1.71
-bleach==3.1.0
-certifi==2018.4.16
-chardet==3.0.4
-click==6.7
-Deprecated==1.2.3
-docutils==0.14
-enum34==1.1.6
-fancycompleter==0.8
-ffmpeg==1.4
-Flask==1.0.2
-idna==2.6
-imagesize==1.1.0
-itsdangerous==0.24
-Jinja2==2.10
-MarkupSafe==1.0
-more-itertools==4.1.0
-numpy==1.14.3
-packaging==19.0
-pandas==0.23.0
-pbr==5.1.3
-pdbpp==0.9.2
-Pillow==5.1.0
-pkginfo==1.5.0.1
-pluggy==0.6.0
-py==1.5.3
-Pygments==2.2.0
-pyparsing==2.3.1
-pytest==3.6.0
-python-dateutil==2.7.3
-pytorch-ignite==0.1.0
-pytz==2018.4
-pyzmq==17.0.0
-readme-renderer==24.0
-requests==2.21.0
-requests-toolbelt==0.9.1
-scipy==1.1.0
-six==1.11.0
-snowballstemmer==1.2.1
-Sphinx==1.8.4
-sphinxcontrib-websupport==1.1.0
-SQLAlchemy==1.3.1
-SQLAlchemy-Utils==0.33.5
-torch>=1.0.1.post2
-torchfile==0.1.0
-tornado==5.0.2
-tqdm==4.24.0
-twine==1.13.0
-urllib3==1.24.1
-visdom==0.1.8.4
-webencodings==0.5.1
-websocket-client==0.47.0
-Werkzeug==0.14.1
-wmctrl==0.3
-wrapt==1.10.11
