+++ paper/paper.md
@@ -20,16 +20,18 @@ bibliography: paper.bib
 
 # Summary
 
-Here, we present a batch-processing library that facilitates pre- and post-processing operations common to machine learning pipelines. Steps in a pipeline are represented using objects that can be composed in a reusable manner with a standard input and output. These input/outputs are performed via a DataFrame-like object called a Message which can have PyTorch [@paszke2017automatic] Tensor-valued columns in addition to all of the functionality of a Pandas DataFrame [@mckinney-proc-scipy-2010]. Each of these pipeline objects can be serialized, enabling one to save checkpoints of the entire state of their workflow rather than just an individual model's. Additionally, we provide a suite of extensions and tools built on top of this library to facilitate common tasks such as relational database access, model training, hyperparameter optimization, saving experimental runs, and more.
+Here, we present a batch-processing library for constructing machine learning pipelines using PyTorch and dataframes. It is meant to provide an easy method to stream data from a dataset into a machine learning model while performing reprocessing steps such as randomization, train/test split, batch normalization, etc. along the way. Fireworks offers more flexibility and structure for constructing input pipelines than the built-in dataset modules in PyTorch [@paszke2017automatic], but is also meant to be easier to use than frameworks such as Apache Spark [@spark].
 
-A pain point in all deep learning frameworks is data entry; one has to take data in its original form and convert it into the form that the model expects (tensors, TFRecords [@tensorflow2015-whitepaper], etc.). The data conversion process can include acquiring the data from a database or an API call, formatting it, applying preprocessing transforms, and preparing mini batches for training. In each of these steps, one must be cognizant of memory, storage, bandwidth, and compute limitations. For example, if a dataset is too big to fit in memory, then it must be streamed or broken into chunks. In practice, dealing with these issues can take more time than developing the actual models. Moreover, ad-hoc solutions to these challenges often introduce biases. For example, if one chooses not to shuffle their dataset before sampling it due to its size, then the order of elements in that dataset becomes a source of bias. Ideally, proper statistical methodology should not have to be compromised for performance and memory considerations, but in practice, that often becomes the case.
+Steps in a pipeline are represented using objects that can be composed in a reusable manner with a standard input and output. These input/outputs are performed via a DataFrame-like object called a Message which can have PyTorch Tensor-valued columns in addition to all of the functionality of a Pandas DataFrame [@mckinney-proc-scipy-2010]. Each of these pipeline objects can be serialized, enabling one to save the entire state of their workflow rather than just an individual model’s when creating checkpoints. Additionally, we provide a suite of extensions and tools built on top of this library to facilitate common tasks such as relational database access, model training, hyperparameter optimization, saving/loading of pipelines, and more. A pain point in all deep learning frameworks is data entry; one has to take data in its original form and convert it into the form that the model expects (for example tensors, TFRecords [@tensorflow2015-whitepaper]). Data conversion can include acquiring the data from a database or an API call, formatting data, applying preprocessing transforms, and preparing mini batches of data for training. At each step, one must be cognizant of memory, storage, bandwidth, and compute limitations. For example, if a dataset is too large to fit in memory, then it must be streamed or broken into chunks. In practice, dealing with these issues is time consuming. Moreover, ad-hoc solutions often introduce biases. For example, if one chooses not to shuffle their dataset before sampling it due to its size, then the order of elements in that dataset becomes a source of bias. Ideally, proper statistical methodology should not have to be compromised for performance and memory considerations, but in practice, that often becomes the case.
 
[…]
