### 2019-06-26 6dfd604
Added JOSS badge
