### 2019-06-27 6e19eea
Converted examples to .ipynb and updated docs.
The MNIST example now has a section in the docs as well.
