+++ paper/paper.md
@@ -31,7 +31,7 @@ These method calls are then recursively chained up the pipeline. So for example,
 
 Because of this modular architecture and the standardization of input/output via the Message object, one can adapt traditional machine learning pipelines built around the Pandas library, because Messages behave like DataFrames. As a result, Fireworks is useful for any data processing task, not just for deep learning pipelines, as described here. It can be used, for example, to interact with a database and construct statistical models with Pandas as well. Additionally, we provides modules for database querying, saving/loading checkpoints, hyperparameter optimization, and model training in addition premade functions for common preprocessing tasks such as batch normalization and one-hot vector mapping using Pipes and Messages. All of these features are designed to improve the productivity of researchers and the reproducibility of their work by reducing the amount of time spent on these tasks and providing a standardized means to do so. The overall design of Fireworks is aimed at transparency and backwards-compatibility with respect to the underlying libraries. This library is designed to be easy to incorporate into a project without requiring one to change the other aspects of their workflow in order to help researchers implement cleaner machine learning pipelines in a reproducible manner to enhance the scientific rigor of their work.
 
-We are currently using this library in our lab to implement neural network models for studying the human microbiome and for finding phage genes within bacterial genomes. In the future, we will implement integration with other popular data science libraries in order to facilitate the inclusion of Fireworks in common data workflows. For example, distributed data processing systems such as Apache Spark could be used to acquire initial data from a larger data-set that can then be pre-processed and mini-batched using Fireworks to train a PyTorch Model. The popular python library SciKit-Learn [@scikit-learn] also has numerous pre-processing tools and models that could be used within a pipeline. There is also a growing interest in conducting data science research on the cloud in a distributed environment. Libraries such as KubeFlow [@kubeflow_2019] enable one to spawn experiments onto a cloud environment so that multiple runs can be performed and tracked simultaneously. In the future, we will integrate these libraries and features with Fireworks.
+We currently use this library in our lab to implement neural network models for studying the human microbiome and for finding phage genes within bacterial genomes. In the future, we will implement integration with other popular data science libraries in order to facilitate the inclusion of Fireworks in common data workflows. For example, distributed data processing systems such as Apache Spark could be used to acquire initial data from a larger data-set that can then be pre-processed and mini-batched using Fireworks to train a PyTorch Model. The popular python library SciKit-Learn [@scikit-learn] also has numerous pre-processing tools and models that could be used within a pipeline. There is also a growing interest in conducting data science research on the cloud in a distributed environment. Libraries such as KubeFlow [@kubeflow_2019] enable one to spawn experiments onto a cloud environment so that multiple runs can be performed and tracked simultaneously. In the future, we will integrate these libraries and features with Fireworks.
 
 
 # References
