+++ docs/conf.py
@@ -49,6 +49,7 @@ extensions = [
     'sphinx.ext.viewcode',
     'sphinx.ext.githubpages',
     'sphinx.ext.napoleon',
+    'sphinx.ext.nbsphinx',
 ]
 
 autodoc_member_order = 'bysource'
