+++ docs/Example.rst
@@ -1,6 +1,8 @@
 Examples
 =====================================
 The below examples walk through the scripts in the examples examples directory of the main project on Github.
+All of these examples can be run as Jupyter notebooks on 'Binder <https://mybinder.org/v2/gh/kellylab/Fireworks/master>'_
+
 
 Making a Model
 ------------------------------
@@ -455,3 +457,258 @@ Model Selection With Databases
         print(params_table.all())
         accuracy_table = DBPipe('accuracy', factory.engine)
         print(accuracy_table.all())
+
+Neural Networks with MNIST 
+------------------------------
+
+The examples above used the simple example of nonlinear regression with synthetic data in order to demonstrate features, but we can use these tools to make more complicated models as well. Let's train a convolutional neural network to classify Fashion MNIST images.
+First, we will download the dataset from the torchvision library, set up an experiment, and convert it to Message format for input into our pipeline.
+
+    .. code-block:: python
+
+        env_name = 'mnist_fashion'
+        # vis = visdom.Visdom(env=env_name) # If you have a running Visdom server, you can uncomment this to generate plots.
+        description = "Here, we will train a convolutional neural network on the Fashion MNIST dataset to demonstrate the usage of Fireworks."
+        experiment = Experiment(env_name, description=description)
+        mnist_dir = env.get('MNIST_DIR', './MNIST/')
+        print(mnist_dir)
+
+        # First, we download our dataset and plot one of its elements as an example.
+        mnist = FashionMNIST(mnist_dir, download=True)
+        dataset = Message({'examples': mnist.data, 'labels': mnist.targets})
+        example = dataset['examples'][0]
+        plt.imshow(example)
+        plt.show()
+
+Next, we will perform a series of preprocessing steps before feeding the data to our model. Using the Pipes built into Fireworks already, we can split our dataset into train and test sets, shuffle it during each epoch, specify a minibatch sizing, and move the batches to GPU memory dynamically.
+However, there are some additional operations we need to perform which are specific to this dataset, and we will implement them as FunctionPipes:
+
+    .. code-block:: python
+
+        # Now we construct our training and test sets as a pipeline.
+        train, test = train_test_split(dataset, test=.1)
+
+        # We can compose pipes to create an input pipeline that will shuffle the training set on each iteration and produce minibatches formatted for our image classifier.
+        shuffler = ShufflerPipe(train)
+        minibatcher = BatchingPipe(shuffler, batch_size=100)
+        to_cuda = TensorPipe(minibatcher, columns=['examples', 'labels']) # By default, all columns will be moved to Cuda if possible, but you can explicitly specify which ones as well
+
+        def tensor_to_float(batch, column='examples'):
+            """ This converts the images from bytes to floats which is the data type that torch.nn.Conv2d expects. """
+            batch[column] = batch[column].float()
+            return batch
+        
+        def reshape_batch(batch, column='examples'):
+            """ This reshapes the batch to have an extra dimension corresponding to the input channels so we can apply the torch.nn.Conv2d operation in our model. """
+            shape = batch[column].shape
+            new_shape = torch.Size([shape[0], 1, shape[1], shape[2]])
+            batch[column] = batch[column].reshape(new_shape)
+            return batch
+
+        def normalize_batch(batch, column='examples'):
+            """ Normalizes pixel intensities to fall between 0 and 1. """
+            batch[column] /= 255. 
+            return batch
+
+        to_float = FunctionPipe(input=to_cuda, function=tensor_to_float)
+        normalized = FunctionPipe(input=to_float, function=normalize_batch)
+        training_set = FunctionPipe(input=normalized, function=reshape_batch)
+
[…]
