### 2019-06-27 65c92d1
Update conf.py
nbsphinx caused an error on readthedocs when building
