+++ docs/conf.py
@@ -49,7 +49,6 @@ extensions = [
     'sphinx.ext.viewcode',
     'sphinx.ext.githubpages',
     'sphinx.ext.napoleon',
-    'sphinx.ext.nbsphinx',
 ]
 
 autodoc_member_order = 'bysource'
