### 2019-06-27 251974c
Development (#74)
* Update conf.py

nbsphinx caused an error on readthedocs when building

* fix typo in docs
