### 2019-07-03 #1478 @arokem
@openjournals/joss-eics : does anyone know what this error means? I am failing to debug this just by looking at the markdown.
