### 2019-07-03 #1478 @kyleniemeyer
I would try removing the empty `orcid:` field for the second author, I think whedon gets confused by that
