### 2019-07-04 3546fa1
Update paper.md
