### 2019-07-07 #1478 @arokem
@smk508 : I implemented @kyleniemeyer's suggestion here: https://github.com/kellylab/Fireworks/pull/75. If you merge that, we can take a look and see what remains to complete the review of this paper.
