### 2019-07-08 #1478 @kyleniemeyer
Ah, it looks like the `affiliations` block is not right. Can you change
```affiliations:
    - name: Systems & Computational Biology, Albert Einstein College of Medicine
```
to
```
affiliations:
    - name: Systems & Computational Biology, Albert Einstein College of Medicine
      index: 1
```
