### 2019-07-15 #1478 @arokem
Sorry for the slowness here. One more comment: the figure caption contains the words "Figure 1". Please remove these, so that you don't get a duplication of these words in the PDF.

Also, the reference to the Paszke paper seems partial. Is there more information?
