### 2019-07-16 89e1db3
Update paper.md
