+++ paper/paper.md
@@ -27,7 +27,7 @@ Steps in a pipeline are represented using objects that can be composed in a reus
 We developed Fireworks to to enable reproducible machine learning pipelines by overcoming some of these challenges in data handling and processing. Pipelines are made up of objects called Pipes which can represent some transformation. Each Pipe has an input and an output. As data flows from Pipe to Pipe, transforms are applied one at a time, and because each Pipe is independent, they can be stacked and reused. Operations are performed on a pipeline by performing method calls on the most downstream Pipe.
 These method calls are then recursively chained up the pipeline. So for example, if one calls iterator methods on the most downstream Pipe (‘iter’ and ‘next’), the iteration will trigger a cascade by which the most upstream Pipe produces its next output and each successive Pipe applies a transformation before handing it off. The same phenomena will apply when calling any method, such as ‘getitem’ for data access. This recursive architecture enables each Pipe to abstract away the inner workings of its upstream Pipe and present a view of the dataset that a downstream Pipe can build additional abstractions off of. For example, a Pipe could cache data from its upstream sources so that whenever a downstream data access call is made, the caching Pipe can retrieve the item from its cache. As another example, a Pipe could virtually shuffle its inputs so that every time an element at a given index is requested, a random element at some other index is returned instead. This framework also enables just-in-time evaluation of  data transformations. A downstream Pipe could call a method only implemented by an upstream Pipe, enabling one to delay applying a transformation until a certain point in a pipeline. Lastly, there are additional primitives for creating branching pipelines and representing PyTorch models as Pipes that are described in the documentation.
 
-![Figure 1: Illustration of the core primitives in Fireworks](Models.png)
+![Illustration of the core primitives in Fireworks.](Models.png)
 
 Because of this modular architecture and the standardization of input/output via the Message object, one can adapt traditional machine learning pipelines built around the Pandas library, because Messages behave like DataFrames. As a result, Fireworks is useful for any data processing task, not just for deep learning pipelines, as described here. It can be used, for example, to interact with a database and construct statistical models with Pandas as well. Additionally, we provides modules for database querying, saving/loading checkpoints, hyperparameter optimization, and model training in addition premade functions for common preprocessing tasks such as batch normalization and one-hot vector mapping using Pipes and Messages. All of these features are designed to improve the productivity of researchers and the reproducibility of their work by reducing the amount of time spent on these tasks and providing a standardized means to do so. The overall design of Fireworks is aimed at transparency and backwards-compatibility with respect to the underlying libraries. This library is designed to be easy to incorporate into a project without requiring one to change the other aspects of their workflow in order to help researchers implement cleaner machine learning pipelines in a reproducible manner to enhance the scientific rigor of their work.
 
