### 2019-07-16 #1478 @smk508
Ok I've updated the figure caption. The Paszke reference was meant to be a reference to the Pytorch library itself and its built-in data processing modules when I was comparing them to Fireworks.
