### 2019-07-16 #1478 @arokem
Gotcha. Looks like it's missing a couple of words, though.

See: https://github.com/pytorch/pytorch/blob/master/CITATION
