+++ paper/paper.bib
@@ -1,10 +1,9 @@
 @inproceedings{paszke2017automatic,
-  title={Automatic differentiation in PyTorch},
+  title={Automatic Differentiation in {PyTorch}},
   author={Paszke, Adam and Gross, Sam and Chintala, Soumith and Chanan, Gregory and Yang, Edward and DeVito, Zachary and Lin, Zeming and Desmaison, Alban and Antiga, Luca and Lerer, Adam},
-  booktitle={NIPS-W},
+  booktitle={NIPS Autodiff Workshop},
   year={2017}
 }
-
 @misc{tensorflow2015-whitepaper,
 title={ {TensorFlow}: Large-Scale machine learning on heterogeneous systems},
 url={https://www.tensorflow.org/},
