### 2019-07-16 #1478 @smk508
Oh sorry. I misunderstood what you were saying before! I fixed the typo in the citations now.
