### 2019-07-16 #1478 @arokem
I think that Whedon is checking all the references in your bib file. I believe everything is in order here.
