### 2019-07-16 #1478 @arokem
@smk508 : could you please create an archive of the software (e.g., using https://zenodo.org) and post the link + DOI here?

I apologize in advance for the additional delay, but I am heading out on vacation and will be back heree next week to move this ahead.
