### 2019-07-17 764cf8c
Added zenodo badge and published a release version
