+++ README.md
@@ -1,6 +1,8 @@
 [![Build Status](https://travis-ci.org/kellylab/fireworks.svg?branch=master)](https://travis-ci.org/kellylab/fireworks)
 [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/kellylab/Fireworks/mnistDemo?filepath=examples%2Fmnist.ipynb)
 [![status](http://joss.theoj.org/papers/6156356281b38760548fdb14505d9ba1/status.svg)](http://joss.theoj.org/papers/6156356281b38760548fdb14505d9ba1)
+[![DOI](https://zenodo.org/badge/135873580.svg)](https://zenodo.org/badge/latestdoi/135873580)
+
 
 # Fireworks - PyTorch with DataFrames
 
