### 2019-07-17 #1478 @smk508
Ok that sounds good. I've made an archive on zenodo by following thir instructions and added a badge to the README.
DOI: 10.5281/zenodo.3339548
URL: https://zenodo.org/badge/latestdoi/135873580
