### 2019-07-20 #1478 @arokem
@openjournals/joss-eics : I believe this paper is ready for your final review
