### 2019-07-20 #1478 @danielskatz
Is version 0.3.0 correct? It seems like perhaps this should be 0.3.2?
