### 2019-07-20 #1478 @danielskatz
Also, I don't think Kubeflow is a journal, so in paper.bib:
@misc{kubeflow_2019, title={Kubeflow}, howpublished={https://www.kubeflow.org/}, journal={Kubeflow}, year={2019}, month={Jan}
}
should probably be:
@misc{kubeflow_2019, title={Kubeflow}, howpublished={https://www.kubeflow.org/}, year={2019}, month={Jan}
}
