+++ paper/paper.bib
@@ -235,5 +235,5 @@ year = {2010}
 }
 @misc{pypi, title={PyPI – the Python Package Index}, url={https://pypi.org/}, journal={PyPI}
 }
-@misc{kubeflow_2019, title={Kubeflow}, howpublished={https://www.kubeflow.org/}, journal={Kubeflow}, year={2019}, month={Jan}
+@misc{kubeflow_2019, title={Kubeflow}, howpublished={https://www.kubeflow.org/}, year={2019}, month={Jan}
 }
