### 2019-07-20 #1478 @smk508
Ok I've updated the kubeflow citation in paper.bib. The current version is 0.3.2, yes. I don't know how to change the top post in this thread to say 0.3.2.
