### 2019-07-21 #1478 @danielskatz
sorry, I see a few more small things in the bib - can we try with https://github.com/kellylab/Fireworks/pull/76 ?
