### 2019-07-21 245b5d4
Update paper.bib (#76)
