+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @inproceedings{paszke2017automatic,
   title={Automatic Differentiation in {PyTorch}},
   author={Paszke, Adam and Gross, Sam and Chintala, Soumith and Chanan, Gregory and Yang, Edward and DeVito, Zachary and Lin, Zeming and Desmaison, Alban and Antiga, Luca and Lerer, Adam},
-  booktitle={NIPS Autodiff Workshop},
+  booktitle={NIPS {A}utodiff {W}orkshop},
   year={2017}
 }
 @misc{tensorflow2015-whitepaper,
@@ -78,7 +78,7 @@ author={
 
 @inproceedings{mckinney-proc-scipy-2010,
 author = {McKinney, Wes},
-booktitle = {Proceedings of the 9th python in science conference},
+booktitle = {Proceedings of the 9th {P}ython in {S}cience {C}onference},
 editor = {van der Walt, St{\'{e}}fan and Millman, Jarrod},
 pages = {51--56},
 title = {{Data Structures for Statistical Computing in Python}},
@@ -235,5 +235,5 @@ year = {2010}
 }
 @misc{pypi, title={PyPI – the Python Package Index}, url={https://pypi.org/}, journal={PyPI}
 }
-@misc{kubeflow_2019, title={Kubeflow}, howpublished={https://www.kubeflow.org/}, year={2019}, month={Jan}
+@misc{kubeflow_2019, title={Kubeflow}, howpublished={\url{https://www.kubeflow.org/}}, year={2019}, month={Jan}
 }
